// src/business-metrics/enums/business-metric-type.enum.ts
export enum BusinessMetricType {
  USER_REGISTRATION = 'USER_REGISTRATION',
  USER_LOGIN = 'USER_LOGIN',
  USER_VERIFICATION = 'USER_VERIFICATION',
  PRODUCE_LISTED = 'PRODUCE_LISTED',
  PRODUCE_SOLD = 'PRODUCE_SOLD',
  OFFER_CREATED = 'OFFER_CREATED',
  OFFER_ACCEPTED = 'OFFER_ACCEPTED',
  TRANSACTION_COMPLETED = 'TRANSACTION_COMPLETED',
  INSPECTION_COMPLETED = 'INSPECTION_COMPLETED',
} 